Move commands into own modules.
